agent-engine/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0
crewai>=0.80.0
pymongo>=4.6.0
mongomock>=4.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
